+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "^5.2.6",
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
@@ -11,6 +10,7 @@
     "@iconify/react": "^4.1.1",
     "@mui/base": "^5.0.0-beta.27",
     "@mui/system": "^5.15.0",
+    "@react-icons/all-files": "^4.1.0",
     "@react-oauth/google": "^0.12.1",
     "@testing-library/jest-dom": "^6.1.6",
     "@testing-library/react": "^14.1.2",
