+++ src/components/units/board/detail/boardDetailContents/BoardDetailContents.types.ts
@@ -1,0 +1,6 @@
+export interface IBoardDetailContents {
+  detailedContent: any;
+  comment: any;
+  data : any;
+  profile : any;
+}
