+++ src/components/units/board/detail/boardDetailFooter/BoardDetailFooter.types.ts
@@ -1,0 +1,4 @@
+export interface IDetailFooter {
+  detailedContent : any,
+  comment : any
+}
