--- src/components/units/board/detail/comment/test.ts
@@ -1,32 +1,0 @@
-export const CommentData = [
-  {
-    commentId: 1,
-    DiaryId: 3,
-    UserId: '철민',
-    content: '그림 잘 그리시네',
-    isEdited: false,
-    createdAt: '2023-12-08T13:07:36.016Z',
-    updatedAt: '2023-12-08T13:07:36.016Z',
-    deletedAt: null,
-  },
-  {
-    commentId: 1,
-    DiaryId: 2,
-    UserId: '덕용',
-    content: '좋습니다',
-    isEdited: false,
-    createdAt: '2023-12-08T13:07:36.016Z',
-    updatedAt: '2023-12-08T13:07:36.016Z',
-    deletedAt: null,
-  },
-  {
-    commentId: 2,
-    DiaryId: 2,
-    UserId: '지우',
-    content: '초상화인가요 ??',
-    isEdited: false,
-    createdAt: '2023-12-08T13:07:36.016Z',
-    updatedAt: '2023-12-08T13:07:36.016Z',
-    deletedAt: null,
-  },
-];
