--- src/components/units/main/test.ts
@@ -1,134 +1,0 @@
-export const dayList = [
-  {
-    EmotionStatus: 1,
-    date: '2023. 12. 9.',
-    id: 1,
-    image: '/pig.png',
-    content: '피곤하다',
-    sentence: '개천에서 용 난다.',
-    writer: '주철민',
-  },
-  {
-    EmotionStatus: 2,
-    date: '2023. 12. 06.',
-    id: 2,
-    image: '/rabbit.png',
-    content: '배고프다',
-    sentence: '돌다리도 두들겨 보고 건너라',
-    writer: '한덕용',
-  },
-  null,
-  null,
-  null,
-  {
-    EmotionStatus: 3,
-    date: '2023. 12. 10.',
-    id: 3,
-    image: '/me.png',
-    content: '화이팅이다.',
-    sentence: '모르는 게 약이다',
-    writer: '송지우',
-  },
-  null,
-  null,
-  {
-    EmotionStatus: 4,
-    date: '2023. 12. 13.',
-    id: 4,
-    image: '/happy.png',
-    content: '그림일기 재밌다',
-    sentence: '아는 게 힘이다 !',
-    writer: '이승연',
-  },
-  null,
-  null,
-  null,
-  null,
-  null,
-  null,
-  null,
-  {
-    EmotionStatus: 3,
-    date: '2023. 12. 21.',
-    id: 5,
-    image: '/angry.png',
-    content: '라면은 안성탕면',
-    sentence: '행복하자 ~~',
-    writer: '유재현',
-  },
-  null,
-  null,
-  null,
-  null,
-  null,
-  null,
-  null,
-  null,
-  {
-    EmotionStatus: 3,
-    date: '2023. 12. 30.',
-    id: 6,
-    image: '/sad.png',
-    content: '노래 추천 받아요',
-    sentence: '할 수 있다 !',
-    writer: '이찬영',
-  },
-];
-
-// export const CalendarContainerDiv = styled.div`
-//   font-family: 'Arial', sans-serif;
-//   text-align: center;
-// `;
-
-// export const Test = styled.div`
-//   background: linear-gradient(
-//     192deg,
-//     #ece9f5 -0.99%,
-//     #e8f5f3 51.07%,
-//     #fcf5f5 99.97%
-//   );
-
-//   background-image: url('/background.png');
-//   background-size: cover;
-//   background-position: center;
-//   background-repeat: no-repeat;
-
-//   box-shadow: 0px 4px 2px 0px rgba(80, 53, 166, 0.2),
-//     0px 4px 10px 0px rgba(255, 255, 255, 0.5) inset,
-//     0px 4px 20px 0px rgba(80, 53, 166, 0.2);
-//   backdrop-filter: blur(15px);
-// `;
-
-// export const LeftRightAnimeButton = styled(motion.div)`
-//   opacity: 0;
-//   transition: opacity 0.35s ease-out, x 0.35s ease-out;
-// `;
-
-// export const CalendarTable = styled.table`
-//   width: 403px;
-//   height: 400px;
-//   margin-left: 20px;
-// `;
-
-// export const TableHead = styled.thead`
-//   background-color: transparent;
-// `;
-
-// export const TableRow = styled.tr`
-//   background-color: transparent; /* 배경색 없애기 */
-//   margin-top: 20px;
-//   margin-bottom: 10px;
-// `;
-
-// export const ThCell = styled.th`
-//   color: #5035a6;
-//   font-family: Pretendard;
-//   font-size: 15px;
-//   font-style: normal;
-//   font-weight: 700;
-//   line-height: 40px;
-//   margin: 20px 0px 20px 0px;
-//   height: 50px;
-//   background-color: transparent;
-// `;
-
